+++ src/Models/Brand.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Brand extends Model
+{
+    protected $fillable = [
+        'name',
+    ];
+}