+++ src/Models/Customer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Customer extends Model
+{
+    protected $fillable = [
+        'firstname',
+        'middlename',
+        'lastname',
+        'email',
+        'password',
+    ];
+}