+++ src/Models/Product.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Product extends Model
+{
+    // do something
+}