+++ src/Resources/BrandResourcePages/CreateBrand.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\BrandResourcePages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Stanleykinkelaar\FilamentShop\Resources\BrandResource;
+
+class CreateBrand extends CreateRecord
+{
+    protected static string $resource = BrandResource::class;
+}