+++ src/Resources/BrandResourcePages/EditBrand.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\BrandResourcePages;
+
+use Filament\Resources\Pages\EditRecord;
+use Stanleykinkelaar\FilamentShop\Resources\BrandResource;
+
+class EditBrand extends EditRecord
+{
+    protected static string $resource = BrandResource::class;
+}