+++ src/Resources/BrandResourcePages/ListBrand.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\BrandResourcePages;
+
+use Filament\Resources\Pages\ListRecords;
+use Stanleykinkelaar\FilamentShop\Resources\BrandResource;
+
+class ListBrand extends ListRecords
+{
+    protected static string $resource = BrandResource::class;
+}