+++ src/Resources/CategoryResourcePages/CreateCategory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\CategoryResourcePages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Stanleykinkelaar\FilamentShop\Resources\CategoryResource;
+
+class CreateCategory extends CreateRecord
+{
+    protected static string $resource = CategoryResource::class;
+}