+++ src/Resources/CategoryResourcePages/EditCategory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\CategoryResourcePages;
+
+use Filament\Resources\Pages\EditRecord;
+use Stanleykinkelaar\FilamentShop\Resources\CategoryResource;
+
+class EditCategory extends EditRecord
+{
+    protected static string $resource = CategoryResource::class;
+}