+++ src/Resources/CategoryResourcePages/ListCategory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\CategoryResourcePages;
+
+use Filament\Resources\Pages\ListRecords;
+use Stanleykinkelaar\FilamentShop\Resources\CategoryResource;
+
+class ListCategory extends ListRecords
+{
+    protected static string $resource = CategoryResource::class;
+}