+++ src/Resources/CustomerResourcePages/ListCustomer.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stanleykinkelaar\FilamentShop\Resources\CustomerResourcePages;
+
+use Filament\Resources\Pages\ListRecords;
+use Stanleykinkelaar\FilamentShop\Resources\CustomerResource;
+
+class ListCustomer extends ListRecords
+{
+    protected static string $resource = CustomerResource::class;
+}