+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
 All notable changes to `filament-shop` will be documented in this file.
+
+# V 0.1
+- Setup basics such as migrations, models, resources
+- **TODO:** add constrains (soms nullable)
+-commands:
+  - pa vendor:publish --tag=filament-shop-seeders
+  - pa vendor:publish --tag=filament-shop-migrations
+  - pa filament-shop